requirements.txt
# --- JAX GPU Setup (GTX 970 / Maxwell / CUDA 11) ---
# CUDA 11対応のホイール置き場を指定
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html

jax==0.4.23
jaxlib==0.4.23+cuda11.cudnn86

# --- Critical Constraints (クラッシュ防止) ---
# JAX 0.4.23 は NumPy 2.x および SciPy 1.13 以降と互換性がありません
numpy<2.0.0
scipy<1.13.0

# --- Neural Network & Optimization ---
# JAX 0.4.23 に対応するバージョン
dm-haiku==0.0.11
optax==0.1.9
chex==0.1.85

# --- RL & Game Environment ---
open_spiel>=1.4.0
gymnasium==0.29.1

# --- Checkpointing & MLOps ---
orbax-checkpoint==0.5.3
mlflow==3.9.0
optuna==3.5.0
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cublas-cu11==11.11.3.6
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cuda-cupti-cu11==11.8.87
nvidia-nccl-cu11==2.21.5
nvidia-cusparse-cu11==11.7.5.86
nvidia-cudnn-cu11==8.6.0.163
google-cloud-storage
streamlit
tensorflow-cpu
